+++ pkg/providers/merge/sig_permission.go
@@ -2,13 +2,12 @@ package merge
 
 import (
 	"github.com/google/go-github/v32/github"
-	"github.com/pingcap-incubator/cherry-bot/pkg/providers/provider"
-	providers "github.com/pingcap-incubator/cherry-bot/pkg/providers/provider"
+	operator "github.com/pingcap-incubator/cherry-bot/pkg/operator"
 )
 
 // AutoMergeAllowName define allow name for auto merge
 
 func (m *merge) CanMergeToMaster(repo string, labels []*github.Label, userName string) error {
-	canMergeRoles := []string{providers.ROLE_COMMITTER, provider.ROLE_COLEADER, provider.ROLE_LEADER}
-	return m.provider.HasPermissionToPRWithLables(repo, labels, userName, canMergeRoles)
+	canMergeRoles := []string{operator.ROLE_COMMITTER, operator.ROLE_COLEADER, operator.ROLE_LEADER}
+	return m.provider.Opr.HasPermissionToPRWithLables(repo, labels, userName, canMergeRoles)
 }
